Meetings
Fixed Up Create Meeting
- Now it looks for employees associated with the house and schedules
them all.

Decided Edit Meeting is fundamentally flawed as designed and needs a
do-over to be effective. Currently no way to select a meeting unless you
know the exact house/date/and times. That's rediculous.
Will implement a drop down with future meetings for the house.