Iterim Commit
Fixed 6 Week House Report. Still Slow as dirt, but it's faster then it
was. No real ideas how to speed it up more.

Noticed that all the tables lost their primary keys when converted from
ACCDB to MYSQL. Updated them.
Changed IHSS hours in the shift_client table from a integer to a
float(7) in preparation for the hour + minute to Hours.minutes
conversion.