Hour + Minutes to Hours.minutes Conversion
Committing work so far.
Shift Edit has been converted.

To Do:
Remaining Time visual needs to be fixed
Default Shifts need to be converted.
Both reports must be converted. (on the plus side this should slightly
increase the speed of the reports)

Points of note. ACCESS VBA cannot handle Modulus by a decimal. It always
returns a 0. It's pretty ugly math. And will require the use of
Math.round on all output to make it readable by humans.