Basic Admin Page
- Altered the table name and columns. For whatever reason I named the
table keys and two of the columns key and value.   All of which are
protected names in access and MySQL. I have no idea what I was thinking.
- Created Demo Admin Form where you can add and remove values from the
list of shift exceptions.
- Added a password feature to the admin button. The password is not
encrypted yet. This is just for demonstration purposes. If we move
forward in this direction I will encrypt the password  and store it in
the database. Current password is NCALC  case sensitive.